refactor : mainPage
