+++ src/components/mainTitle/index.tsx
@@ -1,52 +1,45 @@
 import React from 'react';
 import { Fade, Slide } from 'react-awesome-reveal';
 import styled from 'styled-components';
-import SelectCoin from '../selectcoin';
+import ViewportTypography from '../common/ViewportTypography';
+import { GreenColor } from '../../constants';
+import SelectCoinContainer from './selectCoinContainer';
 import pepeImage from '../../assets/images/pepe.png';
 
-const Wrapper = styled.div`
-  position: absolute;
-  top: 25vh;
-  left: 5vw;
-  font-size: 6vw;
-  font-weight: 700;
-  color: #43841f;
-  z-index: 1;
+const S = {
+  Wrapper: styled.div`
+    position: absolute;
+    top: 25vh;
+    left: 5vw;
+    z-index: 1;
+  `,
+  PepeImage: styled.img`
+    transform: scaleX(-1);
+    height: 100vh;
+    margin-left: 60vw;
 
-  /* 모바일 미디어쿼리 */
-  /* @media (max-width: 480px) {
-    top: 10vh;
-    font-size: 8vw;
-  } */
+    @media (max-width: 1439px) {
+      right: -27%;
+    }
+    @media (max-width: 1023px) {
+      display: none;
+    }
+  `,
+};
 
-  // 모바일
-  // @media (max-width: 767px) {
-  // }
-`;
-
-const PepeImage = styled.img`
-  transform: scaleX(-1);
-  height: 100vh;
-  margin-left: 60vw;
-
-  @media (max-width: 1439px) {
-    right: -27%;
-  }
-  @media (max-width: 1023px) {
-    display: none;
-  }
-`;
 function MainTitle(): any {
   return (
     <>
-      <Wrapper>
+      <S.Wrapper>
         <Fade delay={1000} direction="up" triggerOnce>
-          <p>라고 할 때</p>
+          <ViewportTypography size="6" weight="700" color={GreenColor}>
+            라고 할 때
+          </ViewportTypography>
         </Fade>
-        <SelectCoin />
-      </Wrapper>
+        <SelectCoinContainer />
+      </S.Wrapper>
       <Slide triggerOnce direction="right" duration={1000}>
-        <PepeImage src={pepeImage} alt="pepe" />
+        <S.PepeImage src={pepeImage} alt="pepe" />
       </Slide>
     </>
   );
