+++ src/constants.ts
@@ -1,1 +1,2 @@
 export const GreenColor = '#43841f';
+export const BlackColor = '#000000';
