+++ src/pages/MainPage.tsx
@@ -2,10 +2,10 @@ import React from 'react';
 import { useRecoilValue } from 'recoil';
 import styled from 'styled-components';
 import { calculateButtondAtom } from '../atoms';
-import { Calculation } from '../components/Calculation';
-import MainTitle from '../components/MainTitle';
-import Price from '../components/Price';
-import Product from '../components/Product';
+import { Calculation } from '../components/calculation';
+import MainTitle from '../components/mainTitle';
+import Price from '../components/price';
+import Product from '../components/product';
 
 const WhiteContainer = styled.div`
   height: 30vh;
